Move main.js -> src/main.js
